refactor index.js into ./lib
